[svg] text own module, [all] fmt, some clippy
